requirements.txt
keboola.component==1.4.0
keboola.utils
keboola.http-client
mock~=5.0.2
freezegun~=1.2.2
https://github.com/bakobako/dataconf/zipball/main#egg=dataconf
requests==2.28.2
dateparser==1.1.8